code/ch99_final/RiffMates/requirements.txt
Django==5.0rc1
django-ninja==0.22.2
django-htmx==1.17.0
Pillow==9.5.0
